common: fix type mistach in os_extents_common log
__blksize_t is defined as 'long int' on x86_64, while it is 'int'
on aarch64. Compiling error is raised due to argument type mismatch.
Explicit type cast is added to fix this.

+++ src/common/extent_linux.c
@@ -70,7 +70,7 @@ os_extents_common(const char *path, struct extents *exts,
 	}
 
 	if (exts->extents_count == 0) {
-		LOG(10, "%s: block size: %li", path, st.st_blksize);
+		LOG(10, "%s: block size: %li", path, (long int)st.st_blksize);
 		exts->blksize = (uint64_t)st.st_blksize;
 	}
